Minor code cleanup

+++ Src/PlatformHelpers.h
@@ -30,7 +30,7 @@ namespace DirectX
         virtual const char* what() const override
         {
             static char s_str[64] = {};
-            sprintf_s(s_str, "Failure with HRESULT of %08X", result);
+            sprintf_s(s_str, "Failure with HRESULT of %08X", static_cast<unsigned int>(result));
             return s_str;
         }
